Clean up logger
