+++ dwds/debug_extension_mv3/web/background.dart
@@ -18,6 +18,7 @@ import 'lifeline_ports.dart';
 import 'logger.dart';
 import 'messaging.dart';
 import 'storage.dart';
+import 'utils.dart';
 import 'web_api.dart';
 
 const _authSuccessResponse = 'Dart Debug Authentication Success!';
@@ -69,7 +70,7 @@ Future<bool> _authenticateUser(String extensionUrl, int tabId) async {
   final response = await fetchRequest(authUrl);
   final responseBody = response.body ?? '';
   if (!responseBody.contains(_authSuccessResponse)) {
-    debugWarn('Not authenticated: ${response.status} / $responseBody');
+    debugError('Not authenticated: ${response.status} / $responseBody', devOnly: false,);
     _showWarningNotification('Please re-authenticate and try again.');
     await _createTab(authUrl, inNewWindow: false);
     return false;
@@ -128,7 +129,8 @@ void _setDebuggableIcon() {
 }
 
 void _setDefaultIcon() {
-  chrome.action.setIcon(IconInfo(path: 'dart_grey.png'), /*callback*/ null);
+  final iconPath = isDevMode() ? 'dart_dev.png' : 'dart_grey.png';
+  chrome.action.setIcon(IconInfo(path: iconPath), /*callback*/ null);
 }
 
 Future<DebugInfo?> _fetchDebugInfo(int tabId) {
